Update installation.md
The svelte detection rules are not correct - with the new SvelteKit release ('build`). This made me look into code as well as the documentation. And therefor believe there are minor adjustment needed.

I believe the guidance on webDir is better places at the npx cap init section.  The other amendments are intended to make the user more aware of the things she/he can check when things go wrong - or applying one of the lesser supported/known frameworks.