Reorgnize plugin migration guide
I revised the order of modifications to be per file; previously, the changes were scattered regardless of what file they needed to be made in.  Some places had the filename in the diff, while others had it in the leading text.  To make it consistent, I made a header for each file.